Reformatted conversion of solutions
